+++ src/components/metadataFormControls/ModelTransfer/BaseModelTransfer.tsx
@@ -1,4 +1,4 @@
-import { Transfer, TransferProps } from '@dhis2/ui'
+import { Transfer, TransferOptionProps, TransferProps } from '@dhis2/ui'
 import React, { useCallback, useMemo } from 'react'
 import { DisplayableModel } from '../../../types/models'
 
@@ -7,10 +7,15 @@ const toDisplayOption = (model: DisplayableModel) => ({
     label: model.displayName,
 })
 
+type RenderOptionWithFullValue<TModel> = (
+    props: Omit<TransferOptionProps, 'value'> & { value: TModel }
+) => JSX.Element
+
 type OwnProps<TModel> = {
     selected: TModel[]
     available: TModel[]
     onChange: ({ selected }: { selected: TModel[] }) => void
+    renderOption?: RenderOptionWithFullValue<TModel>
 }
 
 export type BaseModelTransferProps<TModel> = Omit<
@@ -24,12 +29,15 @@ export const BaseModelTransfer = <TModel extends DisplayableModel>({
     available,
     selected,
     onChange,
+    renderOption,
     ...transferProps
 }: BaseModelTransferProps<TModel>) => {
     const { allModelsMap, allTransferOptions } = useMemo(() => {
-        const allModels = selected.concat(available)
+        const allModels = available.concat(selected)
         const allModelsMap = new Map(allModels.map((o) => [o.id, o]))
-        const allTransferOptions = allModels.map(toDisplayOption)
+        const allTransferOptions = Array.from(allModelsMap).map(([i, v]) =>
+            toDisplayOption(v)
+        )
         return {
             allModelsMap,
             allTransferOptions,
@@ -56,10 +64,21 @@ export const BaseModelTransfer = <TModel extends DisplayableModel>({
         [onChange, allModelsMap]
     )
 
+    const renderOptionWithFullValue = useMemo(() => {
+        if (!renderOption) {
+            return undefined
+        }
+        return (props: TransferOptionProps) => {
+            const value = allModelsMap.get(props.value) as TModel
+            return renderOption({ ...props, value })
+        }
+    }, [renderOption, allModelsMap])
+
     return (
         <Transfer
             maxSelections={5000}
             {...transferProps}
+            renderOption={renderOptionWithFullValue}
             selected={selectedTransferValues}
             options={allTransferOptions}
             onChange={handleOnChange}