+++ src/pages/dataSetsWip/form/CategoryComboField.tsx
@@ -1,7 +1,13 @@
 import i18n from '@dhis2/d2-i18n'
 import React from 'react'
-import { ModelSingleSelectField } from '../../../components/metadataFormControls/ModelSingleSelect'
+import {
+    ModelSingleSelect,
+    ModelSingleSelectField,
+    ModelSingleSelectProps,
+} from '../../../components/metadataFormControls/ModelSingleSelect'
 import { DEFAULT_CATEGORY_COMBO } from '../../../lib'
+import { PlainResourceQuery } from '../../../types'
+import { DisplayableModel } from '../../../types/models'
 
 const CATEGORY_COMBOS_QUERY = {
     resource: 'categoryCombos',
@@ -10,10 +16,14 @@ const CATEGORY_COMBOS_QUERY = {
     },
 }
 
-const DEFAULT_CATEGORY_SELECT_OPTION = {
+const DEFAULT_CATEGORYCOMBO_SELECT_OPTION = {
     id: DEFAULT_CATEGORY_COMBO.id,
     displayName: DEFAULT_CATEGORY_COMBO.displayName,
 }
+
+const addDefaultCategoryComboTransform = <TCatCombo extends DisplayableModel>(
+    catCombos: TCatCombo[]
+) => [DEFAULT_CATEGORYCOMBO_SELECT_OPTION as TCatCombo, ...catCombos]
 
 export function CategoryComboField() {
     return (
@@ -24,10 +34,39 @@ export function CategoryComboField() {
                 fieldLabel: i18n.t('Category combination'),
             })}
             query={CATEGORY_COMBOS_QUERY}
-            transform={(catCombos) => [
-                DEFAULT_CATEGORY_SELECT_OPTION,
-                ...catCombos,
-            ]}
+            transform={addDefaultCategoryComboTransform}
+        />
+    )
+}
+
+export type CategoryComboSelectProps<TCatCombo extends DisplayableModel> = Omit<
+    ModelSingleSelectProps<TCatCombo>,
+    'query'
+> & {
+    query: Omit<PlainResourceQuery, 'resource'>
+}
+
+const mergeWithDefaultQuery = (
+    query: Omit<PlainResourceQuery, 'resource'>
+) => ({
+    resource: 'categoryCombos',
+    ...query,
+    params: {
+        ...query.params,
+        filter: ['isDefault:eq:false'].concat(query.params?.filter || []),
+    },
+})
+
+export const CategoryComboSelect = <TCatCombo extends DisplayableModel>(
+    props: CategoryComboSelectProps<TCatCombo>
+) => {
+    const query = mergeWithDefaultQuery(props.query)
+    return (
+        <ModelSingleSelect
+            showNoValueOption
+            transform={addDefaultCategoryComboTransform}
+            {...props}
+            query={query}
         />
     )
 }