+++ src/pages/dataSetsWip/form/dataSetFormSchema.ts
@@ -1,3 +1,4 @@
+import i18n from '@dhis2/d2-i18n'
 import { z } from 'zod'
 import {
     DEFAULT_CATEGORY_COMBO,
@@ -6,8 +7,13 @@ import {
 } from '../../../lib'
 import { createFormValidate } from '../../../lib/form/validate'
 
-const { withAttributeValues, identifiable, style, referenceCollection } =
-    modelFormSchemas
+const {
+    withAttributeValues,
+    identifiable,
+    style,
+    referenceCollection,
+    modelReference,
+} = modelFormSchemas
 
 export const dataSetFormSchema = identifiable
     .merge(withAttributeValues)
@@ -15,14 +21,32 @@ export const dataSetFormSchema = identifiable
         id: z.string().optional(),
         code: z.string().trim().optional(),
         description: z.string().trim().max(2000).optional(),
-        style,
-        dataElements: referenceCollection.default([]),
+        style: style.optional(),
+        dataSetElements: z
+            .array(
+                z.object({
+                    dataElement: modelReference,
+                    categoryCombo: modelReference.optional(),
+                })
+            )
+            .default([]),
         categoryCombo: z
             .object({ id: z.string(), displayName: z.string() })
             .default({ ...DEFAULT_CATEGORY_COMBO }),
+        indicators: referenceCollection.default([]),
         periodType: z.string().default('Monthly'),
+        openFuturePeriods: z
+            .number()
+            .int({ message: i18n.t('The number should not have decimals') }),
+        expiryDays: z.number(),
+        openPeriodsAfterCoEndDate: z
+            .number()
+            .int({ message: i18n.t('The number should not have decimals') })
+            .optional(),
     })
 
 export const initialValues = getDefaults(dataSetFormSchema)
+
+export type DataSetFormValues = typeof initialValues
 
 export const validate = createFormValidate(dataSetFormSchema)